+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 * (improvement) Bump dependencies.
+* (feature) Add `ComponentAssetsHelper` to easily generate component preview screenshot URLs.
 
 
 3.5.1